Amounts are Longs

+++ src/main/java/bitiodine/domain/service/TransactionLocalService.java
@@ -6,8 +6,8 @@ import bitiodine.domain.model.Transaction;
 
 public interface TransactionLocalService {
 	Transaction getOrCreateTransaction(String txHash, List<String> txIns, 
-			List<Integer> amountsIn, List<String> txPrevs, List<String> txOuts,
-			List<Integer> amountsOut, String block_hash, Long timestamp);
+			List<Long> amountsIn, List<String> txPrevs, List<String> txOuts,
+			List<Long> amountsOut, String block_hash, Long timestamp);
 
 	Transaction getOrCreateTransaction(String txHash);
